+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.2.0 	21-05-2015 Joris Verbogt
 - Changed parameter types for arrays
+- Added lifecycle event proxy
+
 
 ## 1.1.0	16-04-2015 Joel Oliveira
 - Added Android support to ti.notificare module